Optimized searchable results, and defaulted region back to the US.
Defaults region to US, based off analytics this makes the most sense.
Cleaned up the detect function to in order to reduce extra component
updates. Changed autofill results to only do one major request for all
searchable characters in a region when it mounts, and instead filters
through that list dynamically on entry rather than recreating requests
to the database on each input. The searchable list is updated based on
region change to keep things separate and request bandwidth a little
lower.